pyproject.toml
[project]
name = "fastapigate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ipython>=8.32.0",
]
